requirements.txt
streamlit
feedparser
beautifulsoup4
requests
python-dotenv
openai
chromadb
sentence-transformers
